+++ SOC_Particle/src/mySummary.h
@@ -30,7 +30,7 @@
 #include "fault.h"
 
 // Function prototypes
-void print_all_fault_buffer(const String code, struct Flt_st *sum, const int iflt, const int nflt);
-void reset_all_fault_buffer(const String code, struct Flt_st *sum, const int iflt, const int nflt);
+void print_all_fault_buffer(const String code, struct Flt_st *sum, const uint16_t iflt, const uint16_t nflt);
+void reset_all_fault_buffer(const String code, struct Flt_st *sum, const uint16_t iflt, const uint16_t nflt);
 
 #endif
